We now support a small subset of the language.
With this commit, we can parse chained and mostly complex expressions.
That said, there are *bound* to be bugs. Even then, I'm not going
to progress much further if I don't start trying to get semantics out
of it. Even weird expressions that are normally illegal can be parsed.
This is because of the incredibibly ambiguous nature of HLSL and C.

So, in order to make sure my grammar is correct, I have to make sure
I can take semantics when and where I need to. So for now, parser
development is going to stop. Next step is going to be a type system,
and then semantics from grammar.